requirements.txt
altair==6.0.0
attrs==25.4.0
backports.zstd==1.3.0
blinker==1.9.0
brotli==1.2.0
cachetools==6.2.6
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
feather-format==0.4.1
filelock==3.24.3
fonttools==4.61.1
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
h5py==3.15.1
idna==3.11
inflate64==1.0.4
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
llvmlite==0.46.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
multivolumefile==0.2.3
narwhals==2.17.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.2
packaging==26.0
pandas==2.3.3
pillow==12.1.1
protobuf==6.33.5
psutil==7.2.2
py7zr==1.1.0
pyarrow==23.0.1
pybcj==1.0.7
pycox==0.3.0
pycryptodomex==3.23.0
pydeck==0.9.1
pyparsing==3.3.2
pyppmd==1.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
sklearn-pandas==2.2.0
smmap==5.0.2
streamlit==1.54.0
sympy==1.14.0
tenacity==9.1.4
texttable==1.7.0
threadpoolctl==3.6.0
toml==0.10.2
torch==2.10.0
torchtuples==0.2.2
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
